extract buildPicture helper from ResponsivePicture
Pulls the hash + dims + variants + srcset construction out of
ResponsivePicture into a shared src/lib/picture.ts. Also exports
withFingerprint so callers that just need the version-stamped URL
don't have to reconstruct it. Sets up the upcoming rehype-plugin
migration that needs the same picture shape.
